renaming of package and improving of swager dto description
