+++ src/main/java/ro/ne8/oauth2/authorizationserver/token/FipsTokenEnhancer.java
@@ -5,7 +5,7 @@ import org.springframework.security.oauth2.common.DefaultOAuth2RefreshToken;
 import org.springframework.security.oauth2.common.OAuth2AccessToken;
 import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.security.oauth2.provider.token.TokenEnhancer;
-import ro.ne8.oauth2.authorizationserver.service.util.DRBGGeneratorBCFipsUtil;
+import ro.ne8.oauth2.authorizationserver.services.util.DRBGGeneratorBCFipsUtil;
 
 public class FipsTokenEnhancer implements TokenEnhancer {
